+++ randovania/games/prime2/game_data.py
@@ -9,10 +9,14 @@ from randovania.games.prime2.layout.preset_describer import echoes_format_params
 def _echoes_gui():
     from randovania.games.prime2.gui.preset_settings import prime2_preset_tabs
     from randovania.games.prime2.gui.dialog.echoes_cosmetic_patches_dialog import EchoesCosmeticPatchesDialog
+    from randovania.gui.game_details.teleporter_details_tab import TeleporterDetailsTab
+    from randovania.games.prime2.gui.translator_gate_details_tab import TranslatorGateDetailsTab
+    from randovania.games.prime2.gui.hint_details_tab import HintDetailsTab
 
     return GameGui(
         tab_provider=prime2_preset_tabs,
         cosmetic_dialog=EchoesCosmeticPatchesDialog,
+        spoiler_visualizer=(TeleporterDetailsTab, TranslatorGateDetailsTab, HintDetailsTab),
     )