+++ randovania/gui/game_details/game_details_tab.py
@@ -1,0 +1,20 @@
+from PySide2 import QtWidgets
+
+from randovania.game_description.game_patches import GamePatches
+from randovania.games.game import RandovaniaGame
+from randovania.interface_common.players_configuration import PlayersConfiguration
+from randovania.layout.base.base_configuration import BaseConfiguration
+
+
+class GameDetailsTab:
+    def __init__(self, parent: QtWidgets.QWidget, game: RandovaniaGame):
+        self.game_enum = game
+
+    def widget(self) -> QtWidgets.QWidget:
+        raise NotImplementedError()
+
+    def tab_title(self) -> str:
+        raise NotImplementedError()
+
+    def update_content(self, configuration: BaseConfiguration, patches: GamePatches, players: PlayersConfiguration):
+        raise NotImplementedError()